Add sync schema: feed sync IDs, tombstones, and engine state storage
Adds sync_id and user_modified_at columns to feeds, sync_tombstones and
sync_state tables, and DatabaseManager+Sync accessors including the
SyncedFeed payload and remote-record upsert. Table creation moves to
DatabaseManager+Tables to keep the class body small.
